py: reorganize
Maybe this is somewhat more like how a package should be organized...
